Fixed integration test by splitting into specific Scala 2.10 and 2.11 versions
